aula10.2.1

+++ praticas/praticas3/questao08.c
@@ -1,17 +1,17 @@
 #include <stdio.h> 
 
 int main() { 
-  int numero;
-  int fatorial;
+  long int numero;
+  long int fatorial;
 
   printf("entre com um numero: ");
-  int deu_certo = scanf("%i", &numero);
+  int deu_certo = scanf("%li", &numero);
 
   if (deu_certo && numero >= 0) {
     for(int i = numero; i > 1; i--) {
       fatorial = fatorial * i;
     }
-    printf("o fatorial do %i = %i\n", numero, fatorial);
+    printf("o fatorial do %li = %li\n", numero, fatorial);
   } else {
     printf("numero invalido");
   }
